Create Week 10 DW
try to familiarize yourself with the functions listed here. should be useful for DW-Physics if they didn't change the questions. 

note: your x-axis should be what you can see and compare, and y-axis is to label it a certain type. in this case, you label if the cells are cancer cells or not.